Move default evaluator to internal from testing
